requirements.CPU.txt
fair-esm==0.1.1
torch==1.8.0
onnx==1.8.1
onnxconverter-common==1.6.0
onnxruntime==1.7.0
onnxruntime-tools==1.6.0